+++ OracleByFPCLtd/OracleByFPCLtd.Tests/obj/Debug/net8.0-windows/OracleByFPCLtd.Tests.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("OracleByFPCLtd.Tests")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+57ca612a2c3837bff66b7074502e41fc67205a6a")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+bb8e6486ff9af8e3c6a4ba5ee6a89542db142ced")]
 [assembly: System.Reflection.AssemblyProductAttribute("OracleByFPCLtd.Tests")]
 [assembly: System.Reflection.AssemblyTitleAttribute("OracleByFPCLtd.Tests")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]